Fix regression with closing and then reopening widgets with QGLWidget
Amends 009abcd7b66738bece6cf354776dfb2ef401636b

Calling QWidget::destroy() on SurfaceAboutToBeDestroyed has a bad side
effect on QGLWidget, because QGLWidget is one of the widgets (or in fact
the only widget?) that reimplements QWidgetPrivate::aboutToDestroy() and
resets the OpenGL context in there (except certain conditions etc.).

If now the window with the QGLWidget is shown again, it will crash as
the QGLWidget is in an invalid state, given that nothing leads to
reinitialization upon the show (which would be incorrect anyway since
we cannot just lose the OpenGL context and then create a new one)

So do the simple solution: pass on the destroyWindow flag from
destroy() to the aboutToDestroy() virtual function, so the
reimplementation can decide what to do.

Fixes: QTBUG-86582
Change-Id: I67bb5f2690927ad62fc041793fbea4ebe6509da4
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>