Fix parsing of dates with spaces in day or month names
Some locales have spaces in their names for months and days. This
breaks parsing dates by splitting on spaces, as we were doing for
TextDate. Reworked the TextDate parsers to take account of this
possibility.

Reworked and improved tst_QDate::roundtripGermanLocale(). It didn't
configure the locale, so it was actually testing whatever locale you
happen to run it with, not specifically German.  It's a TextDate test,
in any case, so should be #if-ed suitably. It merely did the
round-trip, with a comment noting that it should produce no warnings.
Changed it to do comparisons. This, with LANG=vi_VT, made it a
reproducer for the issue.

Fixes: QTBUG-88543
Change-Id: Ie0fb8c6d58f98efcdda2e046523172e66f3491b2
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>